Actualización completa del backend con versión final del equipo (Julian Velasco 909661)

+++ src/main/java/com/uniminuto/clinica/entity/Especializacion.java
@@ -9,10 +9,7 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 import lombok.Data;
 
-/**
- *
- * @author lmora
- */
+
 @Data
 @Entity
 @Table(name = "especializacion")
@@ -21,22 +18,22 @@ public class Especializacion implements Serializable {
      * Id serializable.
      */
     private static final long serialVersionUID = 1L;
-    
+
     /**
      * Id.
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
-    
+    private Long id;
+
     @Column(name = "nombre")
     private String nombre;
-   
+
     @Column(name = "descripcion")
     private String descripcion;
-    
+
     @Column(name = "codigo_especializacion")
     private String codigoEspecializacion;
-    
+
 }
